chore: move ts/ to src/
