+++ .eslintignore
@@ -1,1 +1,1 @@
-ts/message/rpc*
+src/message/rpc*
