+++ package.json
@@ -4,21 +4,20 @@
   "description": "A typescript implementation of gossipsub",
   "leadMaintainer": "Cayman Nava <caymannava@gmail.com>",
   "files": [
-    "src",
     "dist"
   ],
   "type": "module",
-  "types": "dist/ts/index.d.ts",
+  "types": "dist/src/index.d.ts",
   "exports": {
     ".": {
-      "import": "./dist/ts/index.js"
+      "import": "./dist/src/index.js"
     }
   },
   "scripts": {
-    "lint": "eslint --ext .ts ts",
+    "lint": "eslint --ext .ts src",
     "release": "aegir release --no-types",
     "build": "tsc -p tsconfig.build.json",
-    "generate": "protons ./ts/message/rpc.proto",
+    "generate": "protons ./src/message/rpc.proto",
     "prepare": "npm run build",
     "pretest": "npm run build",
     "benchmark": "node ./node_modules/.bin/benchmark 'test/benchmark/time-cache.test.js' --local",
