+++ src/score/constants.ts
