+++ src/score/index.ts
