+++ src/utils/buildRawMessage.ts
