+++ src/utils/has-gossip-protocol.ts
