+++ test/2-nodes.spec.ts
@@ -1,8 +1,8 @@
 import { expect } from 'aegir/utils/chai.js'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
-import type { GossipSub } from '../ts/index.js'
+import type { GossipSub } from '../src/index.js'
 import type { Message, SubscriptionChangeData } from '@libp2p/interfaces/pubsub'
-import { FloodsubID, GossipsubIDv11 } from '../ts/constants.js'
+import { FloodsubID, GossipsubIDv11 } from '../src/constants.js'
 import { pEvent } from 'p-event'
 import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
 import defer from 'p-defer'
