+++ test/go-gossipsub.spec.ts
@@ -4,12 +4,12 @@ import pRetry from 'p-retry'
 import type { EventEmitter } from '@libp2p/interfaces/events'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 import { equals as uint8ArrayEquals } from 'uint8arrays/equals'
-import type { GossipSub, GossipsubEvents } from '../ts/index.js'
-import { MessageAcceptance } from '../ts/types.js'
-import { GossipsubD } from '../ts/constants.js'
+import type { GossipSub, GossipsubEvents } from '../src/index.js'
+import { MessageAcceptance } from '../src/types.js'
+import { GossipsubD } from '../src/constants.js'
 import { fastMsgIdFn } from './utils/index.js'
 import type { Message, SubscriptionChangeData } from '@libp2p/interfaces/pubsub'
-import type { RPC } from '../ts/message/rpc.js'
+import type { RPC } from '../src/message/rpc.js'
 import type { ConnectionManagerEvents } from '@libp2p/interfaces/connection-manager'
 import pWaitFor from 'p-wait-for'
 import { Components } from '@libp2p/interfaces/components'
@@ -24,7 +24,7 @@ import {
 import { FloodSub } from '@libp2p/floodsub'
 import { mockNetwork } from '@libp2p/interface-compliance-tests/mocks'
 import { stop } from '@libp2p/interfaces/startable'
-import { TopicScoreParams } from '../ts/score/peer-score-params.js'
+import { TopicScoreParams } from '../src/score/peer-score-params.js'
 
 /**
  * These tests were translated from:
