+++ test/gossip.spec.ts
@@ -1,8 +1,8 @@
 import { expect } from 'aegir/utils/chai.js'
 import sinon, { SinonStubbedInstance } from 'sinon'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
-import { GossipsubDhi } from '../ts/constants.js'
-import type { GossipSub } from '../ts/index.js'
+import { GossipsubDhi } from '../src/constants.js'
+import type { GossipSub } from '../src/index.js'
 import { pEvent } from 'p-event'
 import { connectAllPubSubNodes, createComponentsArray } from './utils/create-pubsub.js'
 import { Components } from '@libp2p/interfaces/components'
