+++ test/mesh.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from 'aegir/utils/chai.js'
 import delay from 'delay'
-import { GossipsubDhi } from '../ts/constants.js'
-import type { GossipSub } from '../ts/index.js'
+import { GossipsubDhi } from '../src/constants.js'
+import type { GossipSub } from '../src/index.js'
 import { Components } from '@libp2p/interfaces/components'
 import { connectAllPubSubNodes, createComponentsArray } from './utils/create-pubsub.js'
 import { stop } from '@libp2p/interfaces/startable'
