+++ test/peer-score.spec.ts
@@ -2,15 +2,15 @@ import sinon from 'sinon'
 import { expect } from 'aegir/utils/chai.js'
 import delay from 'delay'
 import type { ConnectionManager } from '@libp2p/interfaces/connection-manager'
-import { PeerScore, createPeerScoreParams, createTopicScoreParams } from '../ts/score/index.js'
+import { PeerScore, createPeerScoreParams, createTopicScoreParams } from '../src/score/index.js'
 import { getMsgIdStr, makeTestMessage } from './utils/index.js'
-import { RejectReason } from '../ts/types.js'
-import { ScorePenalty } from '../ts/metrics.js'
+import { RejectReason } from '../src/types.js'
+import { ScorePenalty } from '../src/metrics.js'
 import { stubInterface } from 'ts-sinon'
 import { createEd25519PeerId } from '@libp2p/peer-id-factory'
 import { Components } from '@libp2p/interfaces/components'
-import { PeerStats } from '../ts/score/peer-stats.js'
-import type { PeerScoreParams, TopicScoreParams } from '../ts/score/peer-score-params.js'
+import { PeerStats } from '../src/score/peer-stats.js'
+import type { PeerScoreParams, TopicScoreParams } from '../src/score/peer-score-params.js'
 
 const connectionManager = stubInterface<ConnectionManager>()
 connectionManager.getConnections.returns([])
