+++ test/scoreMetrics.spec.ts
@@ -1,7 +1,7 @@
 import type { ConnectionManager } from '@libp2p/interfaces/connection-manager'
-import { computeAllPeersScoreWeights } from '../ts/score/scoreMetrics.js'
-import { createPeerScoreParams, createTopicScoreParams, PeerScore } from '../ts/score/index.js'
-import { ScorePenalty } from '../ts/metrics.js'
+import { computeAllPeersScoreWeights } from '../src/score/scoreMetrics.js'
+import { createPeerScoreParams, createTopicScoreParams, PeerScore } from '../src/score/index.js'
+import { ScorePenalty } from '../src/metrics.js'
 import { expect } from 'aegir/utils/chai.js'
 import { stubInterface } from 'ts-sinon'
 import { createEd25519PeerId } from '@libp2p/peer-id-factory'
