+++ tsconfig.build.json
@@ -1,5 +1,5 @@
 {
-  "include": ["./ts", "./test"],
+  "include": ["./src", "./test"],
   "compilerOptions": {
     "outDir": "./dist",
     "lib": ["es2020", "dom"],
